Últimos cambios para entrega final

+++ src/components/Cart.js
@@ -4,19 +4,13 @@ import { CartContext } from './CartContext';
 import FormatNumber from '../util/FormatNumber';
 import { doc, collection, setDoc, serverTimestamp, updateDoc, increment, addDoc } from "firebase/firestore";
 import db from '../util/firebaseConfig';
+import Swal from 'sweetalert2';
+import { async } from '@firebase/util';
 const Cart = () => {
     const cart = useContext(CartContext);
     
     const orderCreated = () => {
 
-      // borrar el stock de la bd
-      cart.cartList.forEach(async item => {
-        const updateItem = doc(db, "products", item.idItem);
-        await updateDoc(updateItem, {
-          stock: increment(-item.qtyItem)
-        })
-      });
-  
       // el esquema de la orden con los datos
         let order = {
           buyer: {
@@ -33,8 +27,6 @@ const Cart = () => {
             date: serverTimestamp()
         };
 
-        console.log(order)
-
         // crear la coleccion de la orden en la bd
         const orderInFirestore = async () => {
           const sendOrder = doc(collection(db, "orders"));
@@ -42,23 +34,61 @@ const Cart = () => {
           return sendOrder;
         }
 
-        orderInFirestore()
-        .then(result => alert('Orden creada.\n\nID de su orden: ' + result.id + '\n\n' + 'Gracias por su compra!'))
-        .catch(err => console.log(err))
+          orderInFirestore()
+          .then(result => 
+            Swal.fire({
+            title:' Orden creada. Gracias por su compra.',
+            text: 'ID de su orden' + ' ' + result.id,
+            icon: 'success',
+            backdrop: true
+          }))
+          .catch(err => console.log(err))
+  
+          
+          // borrar el stock de la bd
+          cart.cartList.forEach(async item => {
+            const updateItem = doc(db, "products", item.idItem);
+            await updateDoc(updateItem, {
+              stock: increment(-item.qtyItem)
+            })
+          });
+          
+          cart.deleteAll();
+    }
 
-        cart.deleteAll();
+    // Confirmar si quiere elmiinar el carrito
+    const confirmDeleteAll = () => {
+      Swal.fire({
+        title: 'Está seguro que quiere eliminar todo el carrito?',
+        text: "Esta acción es irreversible",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Confirmar!',
+        backdrop: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Eliminados!',
+            'Se eliminaron todos los productos del carrito.',
+            'success'
+            )
+            cart.deleteAll();
+        }
+      })
     }
 
     return(
 
     <div className="cartContain">
-        <div className="cartTitle"><h1>Mi carrito</h1></div>
+        <div className="cartTitle"><h2>Mi carrito</h2></div>
 
         {
           cart.cartList.length > 0
           ? <div className='cartBtns'> 
               <Link to='/'><button className='cartKeepShop'>Seguir comprando</button></Link>
-              <button className='cartDeleteAll' onClick={cart.deleteAll}>Borrar todos los productos</button>
+              <button className='cartDeleteAll' onClick={confirmDeleteAll}>Borrar todos los productos</button>
             </div>
           : <>
               <p>No hay productos</p>
@@ -78,7 +108,7 @@ const Cart = () => {
                   </div>
                   
                   <div className="cartText">
-                      <h2>{item.titleItem}</h2>
+                      <h3>{item.titleItem}</h3>
                       <p> ${item.priceItem} x {item.qtyItem} item/s</p>
                       <p>Total: <FormatNumber number={cart.calcItem(item.priceItem, item.qtyItem)}/></p>
                       <button onClick={() => cart.deleteItem(item.idItem)}>Eliminar producto</button>
@@ -86,7 +116,6 @@ const Cart = () => {
           
               </div>
                             )
-          
         }
             </div>/
             {/* fin cartboxcontain */}
